requirements.txt
anyascii==0.3.1
asgiref==3.5.2
async-generator==1.10
attrs==21.4.0
beautifulsoup4==4.9.3
certifi==2022.5.18.1
cffi==1.15.1
charset-normalizer==2.0.12
cryptography==37.0.4
Django==4.0.5
django-environ==0.9.0
django-filter==21.1
django-modelcluster==6.0
django-permissionedforms==0.1
django-taggit==2.1.0
django-treebeard==4.5.1
djangorestframework==3.13.1
draftjs-exporter==2.1.7
et-xmlfile==1.1.0
h11==0.13.0
html5lib==1.1
idna==3.3
l18n==2021.3
openpyxl==3.0.10
outcome==1.2.0
Pillow==9.1.1
pycparser==2.21
pyOpenSSL==22.0.0
PySocks==1.7.1
python-slugify==6.1.2
pytz==2022.1
requests==2.28.0
selenium==4.3.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tablib==3.2.1
telepath==0.2
text-unidecode==1.3
trio==0.21.0
trio-websocket==0.9.2
urllib3==1.26.9
wagtail==3.0
webencodings==0.5.1
Willow==1.4.1
wsproto==1.1.0
xlrd==2.0.1
XlsxWriter==3.0.3
xlwt==1.3.0
